+++ src/Stempler/src/Compiler/Renderer/HTMLRenderer.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Compiler\Renderer;
@@ -22,9 +15,6 @@ use Spiral\Stempler\Node\NodeInterface;
 
 final class HTMLRenderer implements RendererInterface
 {
-    /**
-     * @inheritDoc
-     */
     public function render(Compiler $compiler, Compiler\Result $result, NodeInterface $node): bool
     {
         switch (true) {
@@ -47,7 +37,7 @@ final class HTMLRenderer implements RendererInterface
      */
     private function tag(Compiler $compiler, Compiler\Result $result, Tag $node): void
     {
-        $result->push(sprintf('<%s', $node->name), $node->getContext());
+        $result->push(\sprintf('<%s', $node->name), $node->getContext());
 
         foreach ($node->attrs as $attr) {
             if (!$attr instanceof Attribute) {
@@ -58,14 +48,14 @@ final class HTMLRenderer implements RendererInterface
             $this->attribute($compiler, $result, $attr);
         }
 
-        $result->push(sprintf('%s>', $node->void ? '/' : ''), null);
+        $result->push(\sprintf('%s>', $node->void ? '/' : ''), null);
 
         foreach ($node->nodes as $child) {
             $compiler->compile($child, $result);
         }
 
         if (!$node->void) {
-            $result->push(sprintf('</%s>', $node->name), null);
+            $result->push(\sprintf('</%s>', $node->name), null);
         }
     }
 
@@ -75,7 +65,7 @@ final class HTMLRenderer implements RendererInterface
             $result->push(' ', null);
             $compiler->compile($node->name, $result);
         } else {
-            $result->push(sprintf(' %s', $node->name), $node->getContext());
+            $result->push(\sprintf(' %s', $node->name), $node->getContext());
         }
 
         $value = $node->value;
@@ -89,13 +79,13 @@ final class HTMLRenderer implements RendererInterface
             return;
         }
 
-        $result->push(sprintf('=%s', $value), $node->getContext());
+        $result->push(\sprintf('=%s', $value), $node->getContext());
     }
 
     private function verbatim(Compiler $compiler, Compiler\Result $result, Verbatim $node): void
     {
         foreach ($node->nodes as $child) {
-            if (is_string($child)) {
+            if (\is_string($child)) {
                 $result->push($child, null);
                 continue;
             }