+++ src/Stempler/src/Directive/AbstractDirective.php
@@ -1,17 +1,9 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Directive;
 
-use ReflectionObject;
 use Spiral\Stempler\Node\Dynamic\Directive;
 
 /**
@@ -19,20 +11,19 @@ use Spiral\Stempler\Node\Dynamic\Directive;
  */
 abstract class AbstractDirective implements DirectiveRendererInterface
 {
-    /** @var ReflectionObject */
-    private $r;
+    private readonly \ReflectionObject $r;
 
     /**
      * AbstractDirective constructor.
      */
     public function __construct()
     {
-        $this->r = new ReflectionObject($this);
+        $this->r = new \ReflectionObject($this);
     }
 
     public function hasDirective(string $name): bool
     {
-        return $this->r->hasMethod('render' . ucfirst($name));
+        return $this->r->hasMethod('render' . \ucfirst($name));
     }
 
     public function render(Directive $directive): ?string
@@ -41,6 +32,6 @@ abstract class AbstractDirective implements DirectiveRendererInterface
             return null;
         }
 
-        return call_user_func([$this, 'render' . ucfirst($directive->name)], $directive);
+        return \call_user_func([$this, 'render' . \ucfirst($directive->name)], $directive);
     }
 }