+++ src/Stempler/src/Node/PHP.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Node;
@@ -23,25 +16,15 @@ final class PHP implements NodeInterface
 
     public const ORIGINAL_BODY = 'PHP_BODY';
 
-    /** @var string */
-    public $content;
-
-    /** @var array @internal */
-    public $tokens;
-
-    /**
-     * @param Context|null $context
-     */
-    public function __construct(string $content, array $tokens, Context $context = null)
-    {
-        $this->content = $content;
-        $this->tokens = $tokens;
+    public function __construct(
+        public string $content,
+        /** @var array @internal */
+        public array $tokens,
+        Context $context = null
+    ) {
         $this->context = $context;
     }
 
-    /**
-     * @inheritDoc
-     */
     public function getIterator(): \Generator
     {
         yield from [];