+++ src/Stempler/src/Node/Traits/ContextTrait.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Node\Traits;
@@ -15,8 +8,7 @@ use Spiral\Stempler\Parser\Context;
 
 trait ContextTrait
 {
-    /** @var Context|null @internal */
-    private $context;
+    private ?Context $context = null;
 
     public function getContext(): ?Context
     {