+++ src/Stempler/src/Transform/BlockClaims.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Transform;
@@ -17,34 +10,24 @@ namespace Spiral\Stempler\Transform;
  */
 final class BlockClaims
 {
-    /** @var array */
-    private $claimed = [];
+    private array $claimed = [];
 
-    /** @var array[] */
-    private $blocks = [];
-
-    public function __construct(array $blocks)
-    {
-        $this->blocks = $blocks;
+    public function __construct(
+        private readonly array $blocks
+    ) {
     }
 
     public function has(string $name): bool
     {
-        return array_key_exists($name, $this->blocks);
+        return \array_key_exists($name, $this->blocks);
     }
 
-    /**
-     * @return mixed|null
-     */
-    public function get(string $name)
+    public function get(string $name): mixed
     {
         return $this->blocks[$name] ?? null;
     }
 
-    /**
-     * @return mixed|null
-     */
-    public function claim(string $name)
+    public function claim(string $name): mixed
     {
         $this->claimed[] = $name;
 
@@ -53,7 +36,7 @@ final class BlockClaims
 
     public function getNames(): array
     {
-        return array_keys($this->blocks);
+        return \array_keys($this->blocks);
     }
 
     public function getClaimed(): array
@@ -63,6 +46,6 @@ final class BlockClaims
 
     public function getUnclaimed(): array
     {
-        return array_diff(array_keys($this->blocks), $this->claimed);
+        return \array_diff(\array_keys($this->blocks), $this->claimed);
     }
 }