+++ src/Stempler/src/Transform/Context/StackContext.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Transform\Context;
@@ -18,12 +11,9 @@ use Spiral\Stempler\VisitorContext;
 
 final class StackContext
 {
-    /** @var VisitorContext */
-    private $ctx;
-
-    private function __construct(VisitorContext $ctx)
-    {
-        $this->ctx = $ctx;
+    private function __construct(
+        private readonly VisitorContext $ctx
+    ) {
     }
 
     public function register(Aggregate $aggregate, int $level = 0): void
@@ -36,9 +26,6 @@ final class StackContext
         $node->setAttribute(self::class, $stacks);
     }
 
-    /**
-     * @param string|null $uniqueID
-     */
     public function push(string $name, Tag $child, string $uniqueID = null): bool
     {
         foreach ($this->getStacks() as $stack) {
@@ -61,9 +48,6 @@ final class StackContext
         return false;
     }
 
-    /**
-     * @param string|null $uniqueID
-     */
     public function prepend(string $name, Tag $child, string $uniqueID = null): bool
     {
         foreach ($this->getStacks() as $stack) {
@@ -77,7 +61,7 @@ final class StackContext
             $stack->uniqueIDs[$uniqueID] = true;
 
             foreach ($child->nodes as $child) {
-                array_unshift($stack->nodes, $child);
+                \array_unshift($stack->nodes, $child);
             }
 
             return true;
@@ -94,7 +78,7 @@ final class StackContext
     public function getStacks(): array
     {
         $stacks = [];
-        foreach (array_reverse($this->ctx->getScope()) as $node) {
+        foreach (\array_reverse($this->ctx->getScope()) as $node) {
             if ($node instanceof AttributedInterface) {
                 foreach ($node->getAttribute(self::class, []) as $stack) {
                     $stacks[] = $stack;
@@ -118,14 +102,14 @@ final class StackContext
             $scope = $this->ctx->getScope();
 
             // looking for the parent node via given nesting level
-            $node = $scope[count($scope) - 2 - $level] ?? $this->ctx->getFirstNode();
+            $node = $scope[\count($scope) - 2 - $level] ?? $this->ctx->getFirstNode();
         }
 
         if (!$node instanceof AttributedInterface) {
             throw new \LogicException(
-                sprintf(
+                \sprintf(
                     'Unable to create import on node without attribute storage (%s)',
-                    is_object($node) ? get_class($node) : gettype($node)
+                    get_debug_type($node)
                 )
             );
         }